requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.1.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
frozenlist==1.4.0
google-auth==2.24.0
google-auth-oauthlib==1.1.0
gspread==5.12.2
idna==3.6
magic-filter==1.0.12
multidict==6.0.4
oauthlib==3.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
typing_extensions==4.8.0
urllib3==2.1.0
yarl==1.9.3
